+++ src/sync-graphql.test.ts
@@ -11,44 +11,37 @@ import { createSchemaContext } from ".";
 import { Context } from "./types";
 import query from "./query";
 import { SyncMutation } from "./__generated__/sync-mutation";
-import { MultiSyncMutation } from "./__generated__/multi-sync-mutation";
 
 export const TEST_WORKSPACE_ADDR = "+test.a123";
 export const TEST_AUTHOR = generateAuthorKeypair("test") as AuthorKeypair;
 export const TEST_SYNC_MUTATION = /* GraphQL */ `
-  mutation SyncMutation($workspace: String!, $pubUrl: String!) {
-    syncWithPub(workspace: $workspace, pubUrl: $pubUrl) {
+  mutation SyncMutation($workspace: String!, $pubUrls: [String!]!) {
+    syncWithPubs(workspace: $workspace, pubUrls: $pubUrls) {
       __typename
-      ... on DetailedSyncSuccess {
-        pushed {
-          ignoredCount
-          rejectedCount
-          acceptedCount
-        }
-        pulled {
-          ignoredCount
-          rejectedCount
-          acceptedCount
-        }
+      ... on WorkspaceNotValidError {
+        reason
       }
-    }
-  }
-`;
-
-export const TEST_MULTI_SYNC_MUTATION = /* GraphQL */ `
-  mutation MultiSyncMutation($workspaces: [SyncInput!]!) {
-    syncMany(workspaces: $workspaces) {
-      __typename
-      ... on DetailedSyncSuccess {
-        pushed {
-          ignoredCount
-          rejectedCount
-          acceptedCount
+      ... on SyncReport {
+        syncedWorkspace {
+          address
         }
-        pulled {
-          ignoredCount
-          rejectedCount
-          acceptedCount
+        pubSyncResults {
+          __typename
+          ... on PubSyncDetails {
+            pubUrl
+          }
+          ... on DetailedSyncSuccess {
+            pushed {
+              ignoredCount
+              rejectedCount
+              acceptedCount
+            }
+            pulled {
+              ignoredCount
+              rejectedCount
+              acceptedCount
+            }
+          }
         }
       }
     }
@@ -175,17 +168,19 @@ test("Syncs", async () => {
     TEST_SYNC_MUTATION,
     {
       workspace: TEST_WORKSPACE_ADDR,
-      pubUrl: "https://test.server/graphql",
+      pubUrls: ["https://test.server/graphql"],
     },
     ctx
   );
 
-  expect(res.data?.syncWithPub.pulled).toEqual({
+  expect(res.data?.syncWithPubs.__typename).toEqual("SyncReport");
+
+  expect(res.data?.syncWithPubs.pubSyncResults[0].pulled).toEqual({
     acceptedCount: 5,
     ignoredCount: 0,
     rejectedCount: 0,
   });
-  expect(res.data?.syncWithPub.pushed).toEqual({
+  expect(res.data?.syncWithPubs.pubSyncResults[0].pushed).toEqual({
     acceptedCount: 1,
     ignoredCount: 0,
     rejectedCount: 0,
@@ -194,78 +189,6 @@ test("Syncs", async () => {
   const storage = ctx.workspaces[0];
 
   expect(ctx.workspaces.length).toBe(1);
-  expect(res.data).toBeDefined();
-  expect(storage.documents().length).toBe(6);
-  expect(storage.documents().map((doc) => doc.path)).toEqual([
-    "/test/1",
-    "/test/2",
-    "/test/3",
-    "/test/4",
-    "/test/trash",
-    "/wiki/door",
-  ]);
-  expect(storage.documents({ path: "/test/4" })[0].content).toEqual(
-    "A local document"
-  );
-});
-
-test("Multi Syncs", async () => {
-  const ctx = createSchemaContext("MEMORY", {
-    workspaceAddresses: [TEST_WORKSPACE_ADDR],
-  });
-
-  await query(
-    /* GraphQL */ `
-      mutation SetMutation(
-        $author: AuthorInput!
-        $document: NewDocumentInput!
-        $workspace: String!
-      ) {
-        set(author: $author, document: $document, workspace: $workspace) {
-          __typename
-          ... on DocumentRejectedError {
-            reason
-          }
-        }
-      }
-    `,
-    {
-      author: generateAuthorKeypair("test") as AuthorKeypair,
-      document: {
-        path: "/test/4",
-        content: "A local document",
-      },
-      workspace: TEST_WORKSPACE_ADDR,
-    },
-    ctx
-  );
-
-  const res = await query<MultiSyncMutation>(
-    TEST_MULTI_SYNC_MUTATION,
-    {
-      workspaces: [
-        {
-          address: TEST_WORKSPACE_ADDR,
-          pubs: ["https://test.server/graphql"],
-        },
-      ],
-    },
-    ctx
-  );
-
-  expect(res.data?.syncMany[0].pulled).toEqual({
-    acceptedCount: 5,
-    ignoredCount: 0,
-    rejectedCount: 0,
-  });
-  expect(res.data?.syncMany[0].pushed).toEqual({
-    acceptedCount: 1,
-    ignoredCount: 0,
-    rejectedCount: 0,
-  });
-
-  const storage = ctx.workspaces[0];
-
   expect(res.data).toBeDefined();
   expect(storage.documents().length).toBe(6);
   expect(storage.documents().map((doc) => doc.path)).toEqual([
@@ -386,7 +309,7 @@ test("also sends deleted posts", async () => {
     TEST_SYNC_MUTATION,
     {
       workspace: TEST_WORKSPACE_ADDR,
-      pubUrl: "https://test.server/graphql",
+      pubUrls: ["https://test.server/graphql"],
     },
     ctx
   );
@@ -415,7 +338,7 @@ test("also sends deleted posts", async () => {
     TEST_SYNC_MUTATION,
     {
       workspace: TEST_WORKSPACE_ADDR,
-      pubUrl: "https://test.server/graphql",
+      pubUrls: ["https://test.server/graphql"],
     },
     ctx
   );
@@ -437,25 +360,10 @@ test("fails gracefully", async () => {
     TEST_SYNC_MUTATION,
     {
       workspace: "bad.1addr",
-      pubUrl: "https://test.server/graphql",
+      pubUrls: ["https://test.server/graphql"],
     },
     ctx
   );
 
-  expect(res.data?.syncWithPub.__typename).toBe("SyncError");
-
-  const multiRes = await query<MultiSyncMutation>(
-    TEST_MULTI_SYNC_MUTATION,
-    {
-      workspaces: [
-        {
-          address: "wrong.2addr",
-          pubs: ["https://test.server/graphql"],
-        },
-      ],
-    },
-    ctx
-  );
-
-  expect(multiRes.data?.syncMany[0].__typename).toBe("SyncError");
+  expect(res.data?.syncWithPubs.__typename).toBe("WorkspaceNotValidError");
 });