+++ src/types/mutations/mutation.ts
@@ -3,9 +3,8 @@ import { setField } from "./set";
 import { Context } from "../../types";
 import { addWorkspaceField } from "./addWorkspace";
 import { removeWorkspaceField } from "./removeWorkspace";
-import { syncWithPubField } from "./syncWithPub";
+import { syncWithPubsField } from "./syncWithPubs";
 import { ingestDocumentsField } from "./ingestDocuments";
-import { syncManyField } from "./syncMany";
 
 export const mutationType = new GraphQLObjectType<{}, Context>({
   name: "Mutation",
@@ -14,8 +13,7 @@ export const mutationType = new GraphQLObjectType<{}, Context>({
     set: setField,
     addWorkspace: addWorkspaceField,
     removeWorkspace: removeWorkspaceField,
-    syncWithPub: syncWithPubField,
-    syncMany: syncManyField,
+    syncWithPubs: syncWithPubsField,
     ingestDocuments: ingestDocumentsField,
   }),
 });